requirements.txt
OpenCDN-client
pycryptodome